integrate jotai to common state manager
