+++ package.json
@@ -20,6 +20,7 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.23.2",
     "ethereum-checksum-address": "^0.0.8",
+    "jotai": "^2.5.0",
     "lerna": "^6.1.0",
     "prettier": "^2.8.1",
     "react": "~17.0.2",
