+++ packages/common-state/package.json
@@ -25,6 +25,7 @@
     "my-package": "^0.0.0"
   },
   "peerDependencies": {
+    "jotai": "^2.5.0",
     "react": "~17.0.2",
     "react-dom": "~17.0.2",
     "react-scripts": "4.0.3"
