+++ packages/common-state/src/class/index.ts
@@ -1,1 +1,2 @@
+export * from './AtomStateBaseStorage'
 export * from './CommonStatesStorage'
