+++ packages/common-state/src/hooks/HookCommonState.spec.ts
@@ -150,11 +150,14 @@ describe.each<CommonStateTestUtil>([
     expectedValueOnError: undefined,
     expectedCacheOnError: 2,
     initHook: () => {
-      const { state: state1, setRefresh: setRefresh1 } = useCommonState<number>(1)
+      const { state: state1, setRefresh: setRefresh1, setState } = useCommonState<number>(1)
       const { state: state2, setRefresh: setRefresh2 } = useSwitchCommonState<typeof state1, number>(state1, {
         withRefreshOnOriginUpdate: false,
       })
 
+      useEffect(() => {
+        setState((prev) => ({ ...prev, isActual: true }))
+      }, [])
       useEffect(() => setRefresh1({ refreshFn: () => testIncrementor.increment() }), [])
       useEffect(() => setRefresh2({ refreshFn: async (origin) => origin + (await testIncrementor2.increment()) }), [])
 
@@ -203,6 +206,7 @@ describe.each<CommonStateTestUtil>([
         expect(result.current.isLoading).toBe(true)
         expect(result.current.error).toBe(undefined)
 
+        await Promise.resolve()
         jest.runAllTimers()
         await waitForNextUpdate()
 
@@ -240,6 +244,7 @@ describe.each<CommonStateTestUtil>([
         expect(result.current.isLoading).toBe(true)
         expect(result.current.error).toBe(undefined)
 
+        await Promise.resolve()
         jest.runAllTimers()
         await waitForNextUpdate()
 
@@ -269,6 +274,7 @@ describe.each<CommonStateTestUtil>([
         expect(result.current.isLoading).toBe(true)
         expect(result.current.error).toBe(undefined)
 
+        await Promise.resolve()
         jest.runAllTimers()
         await waitForNextUpdate()
 
@@ -306,6 +312,7 @@ describe.each<CommonStateTestUtil>([
         expect(result.current.isLoading).toBe(true)
         expect(result.current.error).toBe(undefined)
 
+        await Promise.resolve()
         jest.runAllTimers()
         await waitForNextUpdate()
 
@@ -323,6 +330,7 @@ describe.each<CommonStateTestUtil>([
           result.current.softRefresh()
         })
 
+        await Promise.resolve()
         jest.runAllTimers()
         await waitForNextUpdate()
 
@@ -344,6 +352,7 @@ describe.each<CommonStateTestUtil>([
         expect(result.current.error).toBe(undefined)
 
         act(() => cancel?.())
+        await Promise.resolve()
         jest.runAllTimers()
         await waitForNextUpdate()
 
@@ -361,6 +370,7 @@ describe.each<CommonStateTestUtil>([
           result.current.softRefresh()
         })
 
+        await Promise.resolve()
         jest.runAllTimers()
         await waitForNextUpdate()
 
@@ -381,6 +391,7 @@ describe.each<CommonStateTestUtil>([
         expect(result.current.isLoading).toBe(true)
         expect(result.current.error).toBe(undefined)
 
+        await Promise.resolve()
         jest.runAllTimers()
         await waitForNextUpdate()
