+++ packages/common-state/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './commonStateFactory'
 export * from './createStateContextEnvironment'
+export * from './storageStateUtils'
